requirements.txt
django==2.1
djangorestframework==3.8.2
elasticsearch-dsl>=5.0,<6.0
python-logstash==0.4.6